+++ Automation/CHANGELOG.md
@@ -1,3 +1,6 @@
+# v3.2.0 (started on Januray 1st, 2026):
+* [Feature] Support district signals for stock value and capacity.
+
 # v3.1.2 (December 17th, 2025):
 * [Change] Support game version `v1.0`. Mod version `v3.0.2` is included to the package for the older game version.
 * [Change] Mod dependency removed: TimberAPI.